+++ pydoclint/utils/internal_error.py
