+++ pydoclint/utils/method_type.py
